priceUnit test 코드 작성
