--- src/test/java/Model/DomainTest.java
@@ -1,5 +1,0 @@
-package Model;
-
-public class DomainTest {
-
-}
